Fixed SSL Issue
There seemed to be an issue where if you used `vagrant up` with the new atlas "update" it would cause an SSL error with this box.  Seems like the issue lies in the fact that I am hosting this box without SSL/TLS.  Went ahead and added the config option "box_download_insecure = true" to bypass cert checking.